Fix incorrect integer division in `getRipeningStage`
The division was missing a cast to float, meaning that it was performing integer / integer division.  This was rounding down to zero in many cases and resulting in fruit not getting dropped.

This fixes a bug that caused extremely low yields from fruit trees from the Binnie ExtraTrees mod in some cases.  One manifestation of this was a Forestry Multifarm playing the rustling sound effect as if it were harvesting a ripe fruit, but then no fruit getting produced.

This commit matches a fix which was applied back in 2018 to the upstream Binnie mod as part of its 1.12 rewrite: https://github.com/ForestryMC/Binnie/commit/6294946aa370b7a7afedaee1a42267b2f9abaa6c#diff-71074c44eae6d20707adf60d373bb2eb7bc2512e1c90cb809d1f0cf328c3b6e5R55

It addresses the root cause of the issue reported here: github.com/GTNewHorizons/GT-New-Horizons-Modpack/issues/19446

A previous attempt to address related fruit drop issues in 92211ddb removed the 5x multiplier, but that was a misdiagnosis, and that commit was later reverted here: https://github.com/GTNewHorizons/Binnie/pull/58.  I've checked that the 5x multiplier is correct and mirrors the behavior of upstream Binnie/ExtraTrees as well as upstream Forestry's own `FruitProviderRipening` logic.

+++ src/main/java/binnie/extratrees/genetics/ExtraTreeFruitGene.java
@@ -387,7 +387,7 @@ public enum ExtraTreeFruitGene implements IAlleleFruit, IFruitProvider {
         if (ripeningTime >= ripeningPeriod) {
             return 1.0f;
         }
-        return ripeningTime / ripeningPeriod;
+        return (float) ripeningTime / ripeningPeriod;
     }
 
     @Override
